requirements.txt
asgiref==3.8.1
attrs==25.3.0
azure-functions==1.23.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
clarabel==0.11.1
click==8.2.1
colorama==0.4.6
cryptography==45.0.5
cvxpy==1.7.1
decompyle3==3.9.2
defusedxml==0.7.1
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.10.0
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
h11==0.16.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.1
llvmlite==0.44.0
lxml==6.0.0
MarkupSafe==3.0.2
mssql-django==1.5
numba==0.61.2
numpy==2.2.6
oauthlib==3.3.1
osqp==1.0.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.3.1
patsy==1.0.1
pycparser==2.22
PyJWT==2.9.0
PyMySQL==1.1.1
pynndescent==0.5.13
pyodbc==5.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-rapidjson==1.21
python3-openid==3.2.0
pytrends==4.9.2
pytz==2025.2
requests==2.32.4
requests-oauthlib==2.0.0
scikit-learn==1.7.1
scipy==1.16.0
scs==3.2.7.post2
selenium==4.34.2
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.7.0
sortedcontainers==2.4.0
spark-parser==1.8.9
sqlparse==0.5.3
statsmodels==0.14.5
threadpoolctl==3.6.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
tritonclient==2.59.0
typing_extensions==4.14.1
tzdata==2025.2
umap-learn==0.5.9.post2
urllib3==2.5.0
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
xdis==6.1.5
